requirements.txt
torch
torchvision
matplotlib
seaborn
scikit-learn
pandas
numpy
jupyter
ipykernel
streamlit
pillow